re-add gateway file + fix version
