+++ discord/ext/listening/__init__.py
@@ -8,4 +8,4 @@ from .processing import *
 from .sink import *
 from .voice_client import *
 
-__version__ = "0.0.4"
+__version__ = "0.0.3"
